20230323/spring/src/main/java/com/example/demo/controllers/WelcomeController.java
package com.example.demo.controllers;

import org.springframework.security.access.annotation.Secured;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
public class WelcomeController {
    @GetMapping("/")
    public String home() {
        return "Hello, world!";
    }

    @GetMapping("/admin")
    @Secured("ROLE_ADMIN")
    public String admin() {
        return "Hello, world!";
    }

    @GetMapping("/user")
    @Secured({"ROLE_USER", "ROLE_ADMIN"})
    public String user() {
        return "Hello, world!";
    }

    @GetMapping("/health")
    public String health() {
        return "Healthy\n";
    }
}
